# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task Management API Documentation",
    "version": "1.0.0",
    "description": "This is the API documentation for the Task Management application."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Event": {
        "type": "object",
        "required": [
          "name",
          "date",
          "capacity",
          "status",
          "category"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the event",
            "example": "event_123abc"
          },
          "name": {
            "type": "string",
            "minLength": 3,
            "maxLength": 100,
            "description": "Event title",
            "example": "Tech Conference 2026"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "description": "Event date in ISO format",
            "example": "2026-06-15T09:00:00Z"
          },
          "capacity": {
            "type": "integer",
            "description": "Maximum number of attendees",
            "example": 150
          },
          "registrationCount": {
            "type": "integer",
            "description": "Number of registered participants",
            "example": 0
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "cancelled",
              "completed"
            ],
            "default": "active",
            "description": "Event status"
          },
          "category": {
            "type": "string",
            "enum": [
              "conference",
              "workshop",
              "meetup",
              "seminar",
              "general"
            ],
            "default": "general",
            "description": "Event category"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the event was created",
            "example": "2026-01-01T10:00:00Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the event was last updated",
            "example": "2026-01-05T12:00:00Z"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/events": {
      "post": {
        "summary": "Create a new event",
        "description": "Add a new event to the system",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "date",
                  "capacity",
                  "status",
                  "category"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Tech Conference 2026"
                  },
                  "date": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2026-06-15T09:00:00Z"
                  },
                  "capacity": {
                    "type": "number",
                    "example": 150
                  },
                  "registrationCount": {
                    "type": "number",
                    "example": 0
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "cancelled",
                      "completed"
                    ],
                    "example": "active"
                  },
                  "category": {
                    "type": "string",
                    "enum": [
                      "conference",
                      "workshop",
                      "meetup",
                      "seminar",
                      "general"
                    ],
                    "example": "conference"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Event created"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Event"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Missing required field: name"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid authentication token"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "get": {
        "summary": "Retrieve all events",
        "description": "Get a list of all events",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Events retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Events retrieved"
                    },
                    "count": {
                      "type": "number",
                      "example": 2
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Event"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid authentication token"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/events/{id}": {
      "get": {
        "summary": "Retrieve a single event",
        "description": "Get details of an event by its ID",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Event retrieved"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Event"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid authentication token"
          },
          "404": {
            "description": "Event not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Event not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "summary": "Update an existing event",
        "description": "Modify an existing event by ID",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Updated Tech Conference"
                  },
                  "date": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2026-06-20T09:00:00Z"
                  },
                  "capacity": {
                    "type": "number",
                    "example": 200
                  },
                  "registrationCount": {
                    "type": "number",
                    "example": 5
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "cancelled",
                      "completed"
                    ],
                    "example": "active"
                  },
                  "category": {
                    "type": "string",
                    "enum": [
                      "conference",
                      "workshop",
                      "meetup",
                      "seminar",
                      "general"
                    ],
                    "example": "conference"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Event updated"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Event"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid authentication token"
          },
          "404": {
            "description": "Event not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Event not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Delete an event",
        "description": "Remove an event by its ID",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Event deleted"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid authentication token"
          },
          "404": {
            "description": "Event not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Event not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "tags": []
}